react-router-dom added to project and home route created
