+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.4",
     "react-scripts": "5.0.0",
     "sass": "^1.56.1",
     "web-vitals": "2.1.4"
